// src/app/interfaces/adminInterface.ts
export interface Admin {
    firstName:      string;
    lastName:       string;
    fatherName:     string;
    dateOfBirth:    Date;
    phoneNumber:    string;
    address:        string;
    gender:         string;
    email:          string;
    userName:       string;
    password:       string;
    activitystatus: string;
    name:           string;
    id?:             number;
}